requirements.txt
google-cloud-aiplatform==1.92.0
google-cloud-storage==2.19.0
google-genai==1.14.0
gitpython==3.1.40
python-dotenv==1.1.0
google-adk==1.0.0
pymupdf==1.26.0
gdown==5.2.0